[consensus] moving TxnManager trait into correct file
The TxnManager trait was living in state_replication.rs whereas it makes more sense for it to live in txn_manager.rs
