+++ consensus/src/chained_bft/test_utils/mock_txn_manager.rs
@@ -1,7 +1,7 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-use crate::state_replication::TxnManager;
+use crate::txn_manager::TxnManager;
 use executor::StateComputeResult;
 use failure::Result;
 use futures::{channel::mpsc, future, Future, FutureExt, SinkExt};
