+++ consensus/src/state_replication.rs
@@ -1,37 +1,14 @@
 // Copyright (c) The Libra Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
+use crate::txn_manager::TxnManager;
 use consensus_types::block::Block;
 use consensus_types::executed_block::ExecutedBlock;
-use executor::{ExecutedTrees, ProcessedVMOutput, StateComputeResult};
+use executor::{ExecutedTrees, ProcessedVMOutput};
 use failure::Result;
 use futures::Future;
 use libra_types::crypto_proxies::{LedgerInfoWithSignatures, ValidatorChangeEventWithProof};
 use std::{pin::Pin, sync::Arc};
-
-/// Retrieves and updates the status of transactions on demand (e.g., via talking with Mempool)
-pub trait TxnManager: Send + Sync {
-    type Payload;
-
-    /// Brings new transactions to be applied.
-    /// The `exclude_txns` list includes the transactions that are already pending in the
-    /// branch of blocks consensus is trying to extend.
-    fn pull_txns(
-        &self,
-        max_size: u64,
-        exclude_txns: Vec<&Self::Payload>,
-    ) -> Pin<Box<dyn Future<Output = Result<Self::Payload>> + Send>>;
-
-    /// Notifies TxnManager about the payload of the committed block including the state compute
-    /// result, which includes the specifics of what transactions succeeded and failed.
-    fn commit_txns<'a>(
-        &'a self,
-        txns: &Self::Payload,
-        compute_result: &StateComputeResult,
-        // Monotonic timestamp_usecs of committed blocks is used to GC expired transactions.
-        timestamp_usecs: u64,
-    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'a>>;
-}
 
 /// While Consensus is managing proposed blocks, `StateComputer` is managing the results of the
 /// (speculative) execution of their payload.
